requirements.txt
Django==6.0.4
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
Pillow>=10.0
psycopg2-binary>=2.9
whitenoise==6.12.0
gunicorn>=21.2
PyJWT==2.12.1
python-decouple>=3.8
dj-database-url>=2.1.0
